Conflicts:
	Senior Project/SeniorProject/SeniorProject.v11.suo
	Senior Project/SeniorProject/SeniorProject/GameEngine.cpp

Updated Hover Information ++

Updated end turn button - not clickable
unless in play phase.

First player game board does not highlight on
spaces u can't place
units.

Can not place unit's on gold mine
locations.

Added warrior ability images and hover images.

+++ Senior Project/SeniorProject/SeniorProject/ClassAbilityAnimator.cpp
@@ -7,10 +7,10 @@ ClassAbilityAnimator::ClassAbilityAnimator(){
 	scaleX					=		0.0f;
 	scaleY					=		0.0f;
 	type					=		-1;
-	animationRect.top		=		0.0f;
-	animationRect.left		=		0.0f;
-	animationRect.right		=		0.0f;
-	animationRect.bottom	=		0.0f;
+	animationRect.top		=		long(0.0f);
+	animationRect.left		=		long(0.0f);
+	animationRect.right		=		long(0.0f);
+	animationRect.bottom	=		long(0.0f);
 	animationActive			=		false;
 };
